Install and use the image builder to create the image including Luci OpenVPN
By building the imagebuilder and using *that* to create the firmware image,
we can put luci (which we download, not compile) into the squashfs instead
of installing it with opkg. This lets us drop opkg, leaving enough space for
a minimal LuCI, OpenVPN and its web interface, all in one image.

This also
   removes "du" command from busybox, which no one seems to use
   removes USB and button support
   removes wpad, which handles WPA support (my network is open!)
   removes swconfig (because the MR3220 is not a switch)
   removes opkg
and
   adds "df", to see how much flash is free: 100k with no settings, 96 with a few.
   adds readlink, whose absence doesn't seem to break anything but is used
   	in /etc/rc.common
   re-adds shadow password support (so you can change the root password again!)